+++ ros-20190910/alibabacloud_ros20190910/models.py
@@ -5606,12 +5606,14 @@ class GetStackRequest(TeaModel):
     def __init__(
         self,
         client_token: str = None,
+        log_option: str = None,
         output_option: str = None,
         region_id: str = None,
         show_resource_progress: str = None,
         stack_id: str = None,
     ):
         self.client_token = client_token
+        self.log_option = log_option
         self.output_option = output_option
         self.region_id = region_id
         self.show_resource_progress = show_resource_progress
@@ -5628,6 +5630,8 @@ class GetStackRequest(TeaModel):
         result = dict()
         if self.client_token is not None:
             result['ClientToken'] = self.client_token
+        if self.log_option is not None:
+            result['LogOption'] = self.log_option
         if self.output_option is not None:
             result['OutputOption'] = self.output_option
         if self.region_id is not None:
@@ -5642,6 +5646,8 @@ class GetStackRequest(TeaModel):
         m = m or dict()
         if m.get('ClientToken') is not None:
             self.client_token = m.get('ClientToken')
+        if m.get('LogOption') is not None:
+            self.log_option = m.get('LogOption')
         if m.get('OutputOption') is not None:
             self.output_option = m.get('OutputOption')
         if m.get('RegionId') is not None:
@@ -5650,6 +5656,80 @@ class GetStackRequest(TeaModel):
             self.show_resource_progress = m.get('ShowResourceProgress')
         if m.get('StackId') is not None:
             self.stack_id = m.get('StackId')
+        return self
+
+
+class GetStackResponseBodyLogResourceLogsLogs(TeaModel):
+    def __init__(
+        self,
+        content: str = None,
+        keys: List[str] = None,
+    ):
+        self.content = content
+        self.keys = keys
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.content is not None:
+            result['Content'] = self.content
+        if self.keys is not None:
+            result['Keys'] = self.keys
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Content') is not None:
+            self.content = m.get('Content')
+        if m.get('Keys') is not None:
+            self.keys = m.get('Keys')
+        return self
+
+
+class GetStackResponseBodyLogResourceLogs(TeaModel):
+    def __init__(
+        self,
+        logs: List[GetStackResponseBodyLogResourceLogsLogs] = None,
+        resource_name: str = None,
+    ):
+        self.logs = logs
+        self.resource_name = resource_name
+
+    def validate(self):
+        if self.logs:
+            for k in self.logs:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['Logs'] = []
+        if self.logs is not None:
+            for k in self.logs:
+                result['Logs'].append(k.to_map() if k else None)
+        if self.resource_name is not None:
+            result['ResourceName'] = self.resource_name
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.logs = []
+        if m.get('Logs') is not None:
+            for k in m.get('Logs'):
+                temp_model = GetStackResponseBodyLogResourceLogsLogs()
+                self.logs.append(temp_model.from_map(k))
+        if m.get('ResourceName') is not None:
+            self.resource_name = m.get('ResourceName')
         return self
 
 
@@ -5695,11 +5775,17 @@ class GetStackResponseBodyLogTerraformLogs(TeaModel):
 class GetStackResponseBodyLog(TeaModel):
     def __init__(
         self,
+        resource_logs: List[GetStackResponseBodyLogResourceLogs] = None,
         terraform_logs: List[GetStackResponseBodyLogTerraformLogs] = None,
     ):
+        self.resource_logs = resource_logs
         self.terraform_logs = terraform_logs
 
     def validate(self):
+        if self.resource_logs:
+            for k in self.resource_logs:
+                if k:
+                    k.validate()
         if self.terraform_logs:
             for k in self.terraform_logs:
                 if k:
@@ -5711,6 +5797,10 @@ class GetStackResponseBodyLog(TeaModel):
             return _map
 
         result = dict()
+        result['ResourceLogs'] = []
+        if self.resource_logs is not None:
+            for k in self.resource_logs:
+                result['ResourceLogs'].append(k.to_map() if k else None)
         result['TerraformLogs'] = []
         if self.terraform_logs is not None:
             for k in self.terraform_logs:
@@ -5719,6 +5809,11 @@ class GetStackResponseBodyLog(TeaModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        self.resource_logs = []
+        if m.get('ResourceLogs') is not None:
+            for k in m.get('ResourceLogs'):
+                temp_model = GetStackResponseBodyLogResourceLogs()
+                self.resource_logs.append(temp_model.from_map(k))
         self.terraform_logs = []
         if m.get('TerraformLogs') is not None:
             for k in m.get('TerraformLogs'):
@@ -7459,12 +7554,14 @@ class GetStackResourceRequest(TeaModel):
         client_token: str = None,
         logical_resource_id: str = None,
         region_id: str = None,
+        resource_attributes: List[str] = None,
         show_resource_attributes: bool = None,
         stack_id: str = None,
     ):
         self.client_token = client_token
         self.logical_resource_id = logical_resource_id
         self.region_id = region_id
+        self.resource_attributes = resource_attributes
         self.show_resource_attributes = show_resource_attributes
         self.stack_id = stack_id
 
@@ -7483,6 +7580,8 @@ class GetStackResourceRequest(TeaModel):
             result['LogicalResourceId'] = self.logical_resource_id
         if self.region_id is not None:
             result['RegionId'] = self.region_id
+        if self.resource_attributes is not None:
+            result['ResourceAttributes'] = self.resource_attributes
         if self.show_resource_attributes is not None:
             result['ShowResourceAttributes'] = self.show_resource_attributes
         if self.stack_id is not None:
@@ -7497,6 +7596,8 @@ class GetStackResourceRequest(TeaModel):
             self.logical_resource_id = m.get('LogicalResourceId')
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
+        if m.get('ResourceAttributes') is not None:
+            self.resource_attributes = m.get('ResourceAttributes')
         if m.get('ShowResourceAttributes') is not None:
             self.show_resource_attributes = m.get('ShowResourceAttributes')
         if m.get('StackId') is not None:
@@ -16657,6 +16758,39 @@ class ValidateTemplateResponseBodyOutputs(TeaModel):
         return self
 
 
+class ValidateTemplateResponseBodyResourceTypes(TeaModel):
+    def __init__(
+        self,
+        data_sources: List[str] = None,
+        resources: List[str] = None,
+    ):
+        self.data_sources = data_sources
+        self.resources = resources
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.data_sources is not None:
+            result['DataSources'] = self.data_sources
+        if self.resources is not None:
+            result['Resources'] = self.resources
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DataSources') is not None:
+            self.data_sources = m.get('DataSources')
+        if m.get('Resources') is not None:
+            self.resources = m.get('Resources')
+        return self
+
+
 class ValidateTemplateResponseBody(TeaModel):
     def __init__(
         self,
@@ -16664,17 +16798,21 @@ class ValidateTemplateResponseBody(TeaModel):
         outputs: List[ValidateTemplateResponseBodyOutputs] = None,
         parameters: List[Dict[str, Any]] = None,
         request_id: str = None,
+        resource_types: ValidateTemplateResponseBodyResourceTypes = None,
     ):
         self.description = description
         self.outputs = outputs
         self.parameters = parameters
         self.request_id = request_id
+        self.resource_types = resource_types
 
     def validate(self):
         if self.outputs:
             for k in self.outputs:
                 if k:
                     k.validate()
+        if self.resource_types:
+            self.resource_types.validate()
 
     def to_map(self):
         _map = super().to_map()
@@ -16692,6 +16830,8 @@ class ValidateTemplateResponseBody(TeaModel):
             result['Parameters'] = self.parameters
         if self.request_id is not None:
             result['RequestId'] = self.request_id
+        if self.resource_types is not None:
+            result['ResourceTypes'] = self.resource_types.to_map()
         return result
 
     def from_map(self, m: dict = None):
@@ -16707,6 +16847,9 @@ class ValidateTemplateResponseBody(TeaModel):
             self.parameters = m.get('Parameters')
         if m.get('RequestId') is not None:
             self.request_id = m.get('RequestId')
+        if m.get('ResourceTypes') is not None:
+            temp_model = ValidateTemplateResponseBodyResourceTypes()
+            self.resource_types = temp_model.from_map(m['ResourceTypes'])
         return self